fastapi-todo-poetry/pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = ""
authors = ["Harry5174 <harisjaved010@gmail.com>"]
readme = "README.md"
packages = [{include = "todo", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
streamlit = "^1.32.2"
sqlalchemy = "^2.0.28"
psycopg2 = "^2.9.9"
requests = "^2.31.0"
sqlmodel = "^0.0.16"
httpx = "^0.27.0"
pytest = "^8.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
